+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 Date format used DD/MM/YYYY
 
+## v0.2 09.08.2020
+#### Added
+* Basic interface
+* Volume control
+* Mute streams
+
+#### fixed
+* Earrape
+
+
 ## v0.1 09.08.2020
 initial release